requirements.txt
streamlit
pypdf
python-dotenv
groq
plotly
pandas